+++ docs/openvsx.md
@@ -7,12 +7,12 @@ This extension is published on [https://open-vsx.org/](https://open-vsx.org/), m
 - The extension is specifically tested on Cursor to ensure the fullest compatibility and a seamless experience.
 - All features available in Visual Studio Code are supported when using Cursor.
 
-> NOTE: Due to tight coupling with debuggers, this extension depends specifically on `anysphere.cpptools`.
+> NOTE: Due to tight coupling with debuggers, this extension depends specifically on `vadimcn.vscode-lldb`.
 
 ## Installation
 
-To install the extension from OpenVSX, search for it from the Cursor marketplace or visit the [RDE ROS 2 listing](https://open-vsx.org/extension/Ranch-Hand-Robotics/rde-ros-2).
+To install the extension from OpenVSX, search for it from the Open-VSX marketplace or visit the [RDE ROS 2 listing](https://open-vsx.org/extension/Ranch-Hand-Robotics/rde-ros-2).
 
 ## Feedback
 
-If you encounter any issues using the extension on Cursor, please open an issue in the repository. 
+If you encounter any issues using this extension, please open an issue in the repository.
