+++ src/extension.ts
@@ -115,9 +115,9 @@ async function startMcpServer(context: vscode.ExtensionContext): Promise<void> {
         const mcpServerPort = vscode_utils.getExtensionConfiguration().get<number>("mcpServerPort", 3002);
         const serverPath = path.join(extPath, "assets", "scripts", "server.py");
         
-                // Show MCP server information for Cursor users
-        const isCursor = vscode_utils.isRunningInCursor();
-        if (isCursor) {
+                // Show MCP server information for users without MCP support
+        const isLldbInstalled = vscode_utils.isLldbExtensionInstalled();
+        if (!isLldbInstalled) {
             const infoMessage = `ROS 2 MCP Server starting on port ${mcpServerPort}.\n\nTo use MCP features in Cursor, add this server to your .cursor/mcp.json:\n\nServer URL: http://localhost:${mcpServerPort}/sse`;
 
             vscode.window.showInformationMessage(infoMessage, "Copy URL", "Open MCP Config", "Dismiss").then(selection => {
@@ -187,9 +187,9 @@ async function startMcpServer(context: vscode.ExtensionContext): Promise<void> {
         }
 
 
-        // Register MCP server definition provider (only in VS Code, not Cursor until sync'd to 1.101)
-        if (!isCursor) {
-            // Check if the MCP API is available (only in VS Code, not Cursor)
+        // Register MCP server definition provider (only when LLDB extension is not available)
+        if (!isLldbInstalled) {
+            // Check if the MCP API is available (only when LLDB extension is not available)
             if ('lm' in vscode && vscode.lm && 'registerMcpServerDefinitionProvider' in vscode.lm) {
                 try {
                     // Use type assertion to handle the API that might not be available in all environments
@@ -248,12 +248,12 @@ export async function activate(context: vscode.ExtensionContext) {
         throw error;
     }
 
-    // Detect if running in Cursor
-    const isCursor = vscode_utils.isRunningInCursor();
-    if (isCursor) {
-        outputChannel.appendLine("Running in Cursor (Anysphere's editor)");
+    // Detect if LLDB extension is installed
+    const isLldbInstalled = vscode_utils.isLldbExtensionInstalled();
+    if (isLldbInstalled) {
+        outputChannel.appendLine("LLDB extension is installed - C++ debugging with LLDB available");
     } else {
-        outputChannel.appendLine("Running in VS Code");
+        outputChannel.appendLine("LLDB extension not installed - C++ debugging with LLDB not available");
     }
 
     // Activate components when the ROS env is changed.
